Don't recreate GnuPG configuration if it exists

+++ Rakefile
@@ -12,7 +12,10 @@ task :default => :spec
 
 namespace :pgp_keys do
   def init_homedir_if_missing
+    return if Dir.exists?(TMP_PGP_HOME)
+
     FileUtils.mkdir_p(TMP_PGP_HOME)
+
     File.write(File.join(TMP_PGP_HOME, "gpg.conf"), <<~GPGCONF)
       personal-digest-preferences SHA512
     GPGCONF
